move files to lib/parachain
