--- parachain/statement.go
@@ -1,138 +1,0 @@
-package parachain
-
-import (
-	"fmt"
-
-	"github.com/ChainSafe/gossamer/lib/common"
-	"github.com/ChainSafe/gossamer/lib/crypto/sr25519"
-	"github.com/ChainSafe/gossamer/pkg/scale"
-)
-
-// A statement, where the candidate receipt is included in the `Seconded` variant.
-type Statement scale.VaryingDataType
-
-func NewStatement() Statement {
-	vdt := scale.MustNewVaryingDataType(Seconded{}, Valid{})
-	return Statement(vdt)
-}
-
-// Set will set a VaryingDataTypeValue using the underlying VaryingDataType
-func (s *Statement) Set(val scale.VaryingDataTypeValue) (err error) {
-	vdt := scale.VaryingDataType(*s)
-	err = vdt.Set(val)
-	if err != nil {
-		return fmt.Errorf("setting value to varying data type: %w", err)
-	}
-
-	*s = Statement(vdt)
-	return nil
-}
-
-// Value returns the value from the underlying VaryingDataType
-func (s *Statement) Value() (scale.VaryingDataTypeValue, error) {
-	vdt := scale.VaryingDataType(*s)
-	return vdt.Value()
-}
-
-// A statement that a validator seconds a candidate.
-type Seconded CommittedCandidateReceipt
-
-func (s Seconded) Index() uint {
-	return 1
-}
-
-// A statement that a validator has deemed a candidate valid.
-type Valid CandidateHash
-
-func (v Valid) Index() uint {
-	return 2
-}
-
-// A candidate-receipt with commitments directly included.
-type CommittedCandidateReceipt struct {
-	// The descriptor of the candidate.
-	Descriptor CandidateDescriptor `scale:"1"`
-
-	// The commitments of the candidate receipt.
-	Commitments CandidateCommitments `scale:"2"`
-}
-
-// This type makes it easy to enforce that a hash is a candidate hash on the type level.
-type CandidateHash struct {
-	Value common.Hash `scale:"1"`
-}
-
-// CandidateDescriptor is a unique descriptor of the candidate receipt.
-type CandidateDescriptor struct {
-	// The ID of the para this is a candidate for.
-	ParaID uint32 `scale:"1"`
-
-	// RelayParent is the hash of the relay-chain block this should be executed in
-	// the context of.
-	// NOTE: the fact that the hash includes this value means that code depends
-	// on this for deduplication. Removing this field is likely to break things.
-	RelayParent common.Hash `scale:"2"`
-
-	// Collator is the collator's sr25519 public key.
-	Collator CollatorID `scale:"3"`
-
-	// PersistedValidationDataHash is the blake2-256 hash of the persisted validation data. This is extra data derived from
-	// relay-chain state which may vary based on bitfields included before the candidate.
-	// Thus, it cannot be derived entirely from the relay-parent.
-	PersistedValidationDataHash common.Hash `scale:"4"`
-
-	// PovHash is the hash of the `pov-block`.
-	PovHash common.Hash `scale:"5"`
-	// ErasureRoot is the root of a block's erasure encoding Merkle tree.
-	ErasureRoot common.Hash `scale:"6"`
-
-	// Signature on blake2-256 of components of this receipt:
-	// The parachain index, the relay parent, the validation data hash, and the `pov_hash`.
-	// this is basically sr25519::Signature
-	Signature CollatorSignature `scale:"7"`
-
-	// ParaHead is the hash of the para header that is being generated by this candidate.
-	ParaHead common.Hash `scale:"8"`
-	// ValidationCodeHash is the blake2-256 hash of the validation code bytes.
-	ValidationCodeHash ValidationCodeHash `scale:"9"`
-}
-
-// CollatorID represents the public key of a collator
-type CollatorID [sr25519.PublicKeyLength]byte
-
-// CollatorSignature is the signature on a candidate's block data signed by a collator.
-type CollatorSignature [sr25519.SignatureLength]byte
-
-// ValidationCodeHash is the blake2-256 hash of the validation code bytes.
-type ValidationCodeHash common.Hash
-
-// CandidateCommitments are Commitments made in a `CandidateReceipt`. Many of these are outputs of validation.
-type CandidateCommitments struct {
-	// Messages destined to be interpreted by the Relay chain itself.
-	UpwardMessages []UpwardMessage `scale:"1"`
-	// Horizontal messages sent by the parachain.
-	HorizontalMessages []OutboundHrmpMessage `scale:"2"`
-	// New validation code.
-	NewValidationCode *ValidationCode `scale:"3"`
-	// The head-data produced as a result of execution.
-	HeadData headData `scale:"4"`
-	// The number of messages processed from the DMQ.
-	ProcessedDownwardMessages uint32 `scale:"5"`
-	// The mark which specifies the block number up to which all inbound HRMP messages are processed.
-	HrmpWatermark uint32 `scale:"6"`
-}
-
-// UpwardMessage A message from a parachain to its Relay Chain.
-type UpwardMessage []byte
-
-// OutboundHrmpMessage is an HRMP message seen from the perspective of a sender.
-type OutboundHrmpMessage struct {
-	Recipient uint32 `scale:"1"`
-	Data      []byte `scale:"2"`
-}
-
-// ValidationCode is Parachain validation code.
-type ValidationCode []byte
-
-// headData is Parachain head data included in the chain.
-type headData []byte
